WIP: integrate pallet_assets into bridge
