+++ Cargo.toml
@@ -22,8 +22,7 @@ members = [
 	"substrate/frame/utility",
 	"pallets/phala",
 	"pallets/phala/mq-runtime-api",
-	"pallets/bridge",
-	"pallets/bridge_transfer",
+	"pallets/xtransfer",
 	"pallets/parachain-info",
 	"runtime/khala",
 	"runtime/thala",
