--- pallets/bridge_transfer/Cargo.toml
@@ -1,57 +1,0 @@
-[package]
-authors = ["Phala Network"]
-edition = "2018"
-name = "pallet-bridge-transfer"
-version = "0.0.1"
-license = "Apache 2.0"
-homepage = "https://phala.network/"
-repository = "https://github.com/Phala-Network/phala-blockchain"
-
-[dependencies]
-# third-party dependencies
-codec = { package = "parity-scale-codec", version = "2.2", default-features = false, features = ["derive"] }
-scale-info = { version = "1.0", default-features = false, features = ["derive"] }
-
-# primitives
-sp-std = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.13", default-features = false }
-sp-runtime = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.13", default-features = false }
-sp-io = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.13", default-features = false }
-sp-core = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.13", default-features = false }
-sp-arithmetic = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.13", default-features = false }
-
-# frame dependencies
-frame-support = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.13", default-features = false }
-frame-system = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.13", default-features = false }
-frame-benchmarking = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.13", default-features = false, optional = true }
-
-phala-types = { path = "../../crates/phala-types", default-features = false }
-phala-pallets = { path = "../phala", default-features = false }
-pallet-bridge = { path = "../bridge", default-features = false }
-hex = { version = "0.4", default-features = false }
-
-[dev-dependencies]
-hex-literal = "0.3"
-pallet-balances = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.13", default-features = false }
-pallet-timestamp = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.13", default-features = false }
-
-[features]
-default = ["std"]
-runtime-benchmarks = [
-	"frame-benchmarking/runtime-benchmarks",
-	"frame-support/runtime-benchmarks",
-	"frame-system/runtime-benchmarks",
-]
-std = [
-	"codec/std",
-	"scale-info/std",
-	"sp-std/std",
-	"sp-runtime/std",
-	"sp-io/std",
-	"sp-core/std",
-	"sp-arithmetic/std",
-	"frame-support/std",
-	"frame-system/std",
-	"frame-benchmarking/std",
-	"pallet-bridge/std",
-	"phala-pallets/std",
-]
