+++ pallets/xtransfer/src/bridge_transfer/mock.rs
@@ -10,10 +10,9 @@ use sp_runtime::{
 	Perbill,
 };
 
-use crate::{self as bridge_transfer, Config};
+use crate::bridge_transfer;
 pub use pallet_balances as balances;
-use pallet_bridge as bridge;
-use phala_pallets::{pallet_mq as mq, pallet_registry as reg};
+use crate::pallet_bridge as bridge;
 
 type UncheckedExtrinsic = frame_system::mocking::MockUncheckedExtrinsic<Test>;
 type Block = frame_system::mocking::MockBlock<Test>;
@@ -28,9 +27,9 @@ frame_support::construct_runtime!(
 		Balances: pallet_balances::{Pallet, Call, Storage, Config<T>, Event<T>},
 		Bridge: bridge::{Pallet, Call, Storage, Event<T>},
 		BridgeTransfer: bridge_transfer::{Pallet, Call, Storage, Event<T>},
-		PhalaMq: mq::{Pallet, Call, Storage},
-		PhalaRegistry: reg::{Pallet, Call, Event, Storage},
 		Timestamp: pallet_timestamp::{Pallet, Call, Storage, Inherent},
+		Assets: pallet_assets::{Pallet, Call, Storage, Event<T>},
+		AssetsWrapper: pallet_assets_wrapper::{Pallet, Call, Storage, Event<T>},
 	}
 );
 
@@ -111,37 +110,9 @@ impl Config for Test {
 	type Event = Event;
 	type BridgeOrigin = bridge::EnsureBridge<Test>;
 	type Currency = Balances;
+	type Assets = AssetsWrapper;
 	type NativeTokenResourceId = NativeTokenResourceId;
 	type OnFeePay = ();
-}
-
-impl mq::Config for Test {
-	type CallMatcher = MqCallMatcher;
-	type QueueNotifyConfig = ();
-}
-
-pub struct MqCallMatcher;
-impl mq::CallMatcher<Test> for MqCallMatcher {
-	fn match_call(call: &Call) -> Option<&mq::Call<Test>> {
-		match call {
-			Call::PhalaMq(mq_call) => Some(mq_call),
-			_ => None,
-		}
-	}
-}
-
-parameter_types! {
-	pub const VerifyPRuntime: bool = false;
-	pub const VerifyRelaychainGenesisBlockHash: bool = false;
-}
-
-impl reg::Config for Test {
-	type Event = Event;
-	type AttestationValidator = reg::IasValidator;
-	type UnixTime = Timestamp;
-	type VerifyPRuntime = VerifyPRuntime;
-	type VerifyRelaychainGenesisBlockHash = VerifyRelaychainGenesisBlockHash;
-	type GovernanceOrigin = frame_system::EnsureRoot<Self::AccountId>;
 }
 
 impl pallet_timestamp::Config for Test {
