+++ pallets/xtransfer/src/lib.rs
@@ -4,7 +4,11 @@
 pub use crate::xcm::{xcm_helper, xcm_transfer as pallet_xcm_transfer};
 mod xcm;
 
-pub mod assets_wrapper;
+pub mod bridge;
+pub use bridge as pallet_bridge;
 
-// Alias
+pub mod bridge_transfer;
+pub use bridge_transfer as pallet_bridge_transfer;
+
+pub mod assets_wrapper;
 pub use assets_wrapper as pallet_assets_wrapper;
