+++ runtime/thala/Cargo.toml
@@ -84,8 +84,6 @@ pallet-xcm = { git = "https://github.com/paritytech/polkadot", branch = "release
 
 # Local dependencies
 pallet-parachain-info = { path = "../../pallets/parachain-info", default-features = false }
-pallet-bridge = { path = "../../pallets/bridge", default-features = false }
-pallet-bridge-transfer = { path = "../../pallets/bridge_transfer", default-features = false }
 phala-pallets = { path = "../../pallets/phala", default-features = false }
 pallet-mq-runtime-api = { path = "../../pallets/phala/mq-runtime-api", default-features = false }
 xtransfer-pallets = { path = "../../pallets/xtransfer", default-features = false }
@@ -121,7 +119,7 @@ runtime-benchmarks = [
 	"pallet-timestamp/runtime-benchmarks",
 	"pallet-tips/runtime-benchmarks",
 	"phala-pallets/runtime-benchmarks",
-	"pallet-bridge/runtime-benchmarks",
+	"xtransfer-pallets/runtime-benchmarks",
 	"pallet-assets/runtime-benchmarks"
 ]
 std = [
@@ -185,8 +183,6 @@ std = [
 	"pallet-tips/std",
 	"pallet-assets/std",
 	"pallet-parachain-info/std",
-	"pallet-bridge/std",
-	"pallet-bridge-transfer/std",
 	"phala-pallets/std",
 	"pallet-mq-runtime-api/std",
 	"xtransfer-pallets/std",
