+++ runtime/thala/src/lib.rs
@@ -99,7 +99,7 @@ pub use parachains_common::Index;
 
 pub use phala_pallets::{pallet_mining, pallet_mq, pallet_registry, pallet_stakepool};
 
-pub use xtransfer_pallets::{pallet_assets_wrapper, pallet_xcm_transfer, xcm_helper};
+pub use xtransfer_pallets::{pallet_assets_wrapper, pallet_bridge, pallet_bridge_transfer, pallet_xcm_transfer, xcm_helper};
 pub use pallet_assets_wrapper::XTransferAssetId;
 
 #[cfg(any(feature = "std", test))]
@@ -1261,6 +1261,7 @@ impl pallet_bridge_transfer::Config for Runtime {
     type Event = Event;
     type BridgeOrigin = pallet_bridge::EnsureBridge<Runtime>;
     type Currency = Balances;
+    type Assets = AssetsWrapper;
     type NativeTokenResourceId = NativeTokenResourceId;
     type OnFeePay = Treasury;
 }
